requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cached-property==1.5.2
certifi==2022.12.7
certstream==1.12
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dnspython==2.3.0
idna==3.4
importlib-resources==5.12.0
isodate==0.6.1
jsonschema==4.17.3
lxml==4.9.2
oauthlib==3.2.2
pdfminer.six==20201018
pdfx==1.4.1
pkgutil_resolve_name==1.3.10
publicsuffixlist==0.9.4
pycparser==2.21
Pygments==2.15.1
PyMySQL==1.1.0
pyrsistent==0.19.3
pyspnego==0.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.4.1
termcolor==2.3.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
websocket-client==1.6.3
wrapt==1.15.0
yara-python==4.3.1
zipp==3.15.0
